-various network fixes
git-svn-id: file:///media/data/other/svn/repos/BountyBandits/trunk@190 f34f5d6a-d8fa-4a85-8d7d-3577574e89da

+++ BountyBandits/Network/Structs.cs
@@ -60,7 +60,7 @@ namespace BountyBandits.Network
             msg.Write(gameItem.body.LinearVelocity.Y);
             msg.Write(gameItem.body.Rotation);
             msg.Write(gameItem.body.AngularVelocity);
-            msg.Write((Int16)(gameItem.body.Tag == null ? 0 : (int)gameItem.body.Tag));
+            msg.Write((Int16)(gameItem.body.Tag == null ? (short)0 : gameItem.body.Tag));
             msg.Write(gameItem.guid.ToString());
         }
